move net sensor to others
